+++ ios/Classes/IntercomFlutterPlugin.m
@@ -1,6 +1,23 @@
 #import "IntercomFlutterPlugin.h"
-#import "Intercom.h"
+#import <Intercom/Intercom.h>
 #import <UserNotifications/UserNotifications.h>
+
+id unread;
+
+@implementation UnreadStreamHandler
+- (FlutterError*)onListenWithArguments:(id)arguments eventSink:(FlutterEventSink)eventSink {
+    unread = [[NSNotificationCenter defaultCenter] addObserverForName:IntercomUnreadConversationCountDidChangeNotification object:nil queue:[NSOperationQueue mainQueue] usingBlock:^(NSNotification * _Nonnull note) {
+        NSNumber *myNum = @([Intercom unreadConversationCount]);
+        eventSink(myNum);
+    }];
+  return nil;
+}
+
+- (FlutterError*)onCancelWithArguments:(id)arguments {
+    [[NSNotificationCenter defaultCenter] removeObserver:unread];
+  return nil;
+}
+@end
 
 @implementation IntercomFlutterPlugin
  FlutterMethodChannel *_channel;
@@ -12,11 +29,17 @@
     id instance = [[IntercomFlutterPlugin alloc] initWithChannel:channel];
     [registrar addApplicationDelegate:instance];
     [registrar addMethodCallDelegate:instance channel:channel];
+    FlutterEventChannel* unreadChannel = [FlutterEventChannel eventChannelWithName:@"maido.io/intercom/unread"
+    binaryMessenger:[registrar messenger]];
+    UnreadStreamHandler* unreadStreamHandler =
+        [[UnreadStreamHandler alloc] init];
+    [unreadChannel setStreamHandler:unreadStreamHandler];
+
 }
 
 - (instancetype)initWithChannel:(FlutterMethodChannel *)channel {
     self = [super init];
-    
+
     if (self) {
         _channel = channel;
     }